フォロワーのリストを完成

+++ assets/js/src/components/follow-toggle.js
@@ -14,8 +14,12 @@
             $btn = $(this),
             following = $(this).hasClass('btn-following');
         e.preventDefault();
+        if( $btn.hasClass('btn-follow--loading') ){
+            return;
+        }
         if( following ){
             Hametuha.confirm('フォローを解除してよろしいですか？', function(){
+                $btn.addClass('btn-follow--loading');
                 $.ajax( {
                     url: endpoint,
                     method: 'DELETE',
@@ -26,10 +30,13 @@
                 } ).done( function ( response ) {
                     $btn.removeClass('btn-following');
                 }).fail(function(){
-                    Hametuha.alert('フォローを解除できませんでした');
+                    Hametuha.alert('フォローを解除できませんでした', true);
+                }).always(function(){
+                   $btn.removeClass('btn-follow--loading');
                 });
-            });
+            }, true);
         }else{
+            $btn.addClass('btn-follow--loading');
             $.ajax( {
                 url: endpoint,
                 method: 'POST',
@@ -40,7 +47,9 @@
             } ).done( function ( response ) {
                 $btn.addClass('btn-following');
             }).fail(function(){
-               Hametuha.alert('フォローに失敗しました。すでにフォロー済みか、サーバが混み合っています。');
+               Hametuha.alert('フォローに失敗しました。すでにフォロー済みか、サーバが混み合っています。', true);
+            }).always(function(){
+                $btn.removeClass('btn-follow--loading');
             });
         }
